backend/machinelearning/requirements.txt
pandas==1.5.1
joblib==1.2.0
lightgbm==3.3.5
numpy==1.24.2
scikit-learn==1.2.2
scipy==1.10.1
threadpoolctl==3.1.0
flask==2.0.1
boto3==1.26.69
python-dotenv==1.0.0
flask_cors==3.0.10